minor fixes to make api compliant

+++ src/worker.c
@@ -38,7 +38,7 @@ worker_thread_params* get_worker_thread_params(sync_queue* job_queue, uint64_t j
 
 static void* worker_function(void* args);
 
-int start_worker(pthread_t* thread_id, sync_queue* job_queue, unsigned long long int job_queue_empty_timeout_in_microseconds, void(*start_up)(void* additional_params), void(*clean_up)(void* additional_params), void* additional_params)
+int start_worker(pthread_t* thread_id, sync_queue* job_queue, uint64_t job_queue_empty_timeout_in_microseconds, void(*start_up)(void* additional_params), void(*clean_up)(void* additional_params), void* additional_params)
 {
 	// default return value is the insufficient resources to create a thread
 	int return_val = EAGAIN;
@@ -61,7 +61,7 @@ int start_worker(pthread_t* thread_id, sync_queue* job_queue, unsigned long long
 	return return_val;
 }
 
-int submit_job_worker(sync_queue* job_queue, void* (*function_p)(void* input_p), void* input_p, promise* promise_for_output, void (*cancellation_callback)(void* input_p), unsigned long long int submission_timeout_in_microseconds)
+int submit_job_worker(sync_queue* job_queue, void* (*function_p)(void* input_p), void* input_p, promise* promise_for_output, void (*cancellation_callback)(void* input_p), uint64_t submission_timeout_in_microseconds)
 {
 	int was_job_queued = 0;
 
@@ -86,7 +86,7 @@ int submit_job_worker(sync_queue* job_queue, void* (*function_p)(void* input_p),
 	return was_job_queued;
 }
 
-int submit_stop_worker(sync_queue* job_queue, unsigned long long int submission_timeout_in_microseconds)
+int submit_stop_worker(sync_queue* job_queue, uint64_t submission_timeout_in_microseconds)
 {
 	return push_sync_queue(job_queue, NULL, submission_timeout_in_microseconds);
 }
